refactor(app): moved some files around and removed any unnecessary files/code
