+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 typings
-app/*.js
-app/*.map
+app/**/*.js
+app/**/*.map
